+++ pkg/container/podman_provider_test.go
@@ -3,6 +3,7 @@
 package container
 
 import (
+	"bytes"
 	"errors"
 	"io"
 	"strings"
@@ -13,14 +14,14 @@ import (
 )
 
 func TestPodmanProviderInfo(t *testing.T) {
-	rawOutput := strings.NewReader(
+	rawOutput := []byte(
 		`{"version": {"version": "2.2.1"}, "store": {"graphDriverName": "overlay", "graphRoot": "/var/lib/containers/storage", "runRoot": "/run/user/0", "volumePath": "/var/lib/containers/storage/volumes"}}`,
 	)
 
 	// Mock executePodmanInfoFunc to return expected output
 	originalFunc := executePodmanInfoFunc
 	executePodmanInfoFunc = func() (stdout, stderr io.Reader, err error) {
-		stdout = rawOutput
+		stdout = bytes.NewReader(rawOutput)
 		return stdout, stderr, err
 	}
 	defer func() {
@@ -64,7 +65,9 @@ func TestPodmanProviderInfo(t *testing.T) {
 	assert.Equal(t, expected, podmanInfo.Results())
 
 	// Check the raw data
-	assert.Equal(t, rawOutput, podmanInfo.RawData())
+	infoOutputBytes, err := io.ReadAll(podmanInfo.RawData())
+	assert.NoError(t, err)
+	assert.Equal(t, rawOutput, infoOutputBytes)
 }
 
 func TestPodmanProviderInfoParseError(t *testing.T) {