+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.4.2] - 2025-03-25
 
+### Added
+- Conjur configuration files are now included in the inspect report. CNJR-8738
+
 ### Fixed
 - Podman Info raw output is now correctly stored in the inspect report archive.